#37 refactor: 커스텀 훅 생성해서 로직 분리
