+++ pages/wikilist/index.tsx
@@ -1,60 +1,14 @@
 import WikiCard from '@/components/WikiList/WikiCard';
-import ProfileType from '@/types/types';
 import styles from './WikiList.module.scss';
-import { getProfileList } from '@/apis/auth/getProfileList';
-import { useCallback, useEffect, useState } from 'react';
 import Image from 'next/image';
 import Spinner from '@/components/WikiList/SPinner';
-import { useInView } from 'react-intersection-observer';
-import { useRouter } from 'next/router';
 import TheLatestSearch from '@/components/WikiList/TheLatestSearch';
 import { useSearch } from '@/contexts/SearchProvider';
+import { useWikiInfiniteScroll } from '@/hooks/WikiList/useWikiInfiniteScroll';
 
 export default function WikiList() {
   const { searchTerm } = useSearch();
-  const router = useRouter();
-  const [profileCards, setProfileCards] = useState<ProfileType[]>([]);
-  const [loading, setLoading] = useState<boolean>(true);
-  const [page, setPage] = useState<number>(1);
-  const [ref, inView] = useInView({ threshold: 0 });
-  const [hasMore, setHasMore] = useState<boolean>(true);
-
-  const handleLoadProfileCards = useCallback(async (page: number, pageSize: number, name: string) => {
-    if (!hasMore) {
-      return;
-    }
-    setLoading(true);
-    try {
-      const nextProfileCards = await getProfileList(page, pageSize, name);
-      setProfileCards(prevCards => [...prevCards, ...nextProfileCards]);
-      setPage(prevPage => prevPage + 1);
-
-      if (nextProfileCards.length < pageSize) {
-        setHasMore(false);
-        setPage(1);
-      }
-    } catch (error) {
-      console.error('Failed to fetch profileCards', error);
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  //searchTerm에 값이 할당되어 있고 router가 준비되었을 때 리스트 초기화
-  useEffect(() => {
-    if (searchTerm !== undefined && router.isReady) {
-      setProfileCards([]);
-      setHasMore(true);
-      handleLoadProfileCards(1, 4, searchTerm);
-    }
-  }, [searchTerm]);
-
-  //데이터 더 불러오기
-  useEffect(() => {
-    if (inView && hasMore && page > 1) {
-      handleLoadProfileCards(page, 4, searchTerm);
-    }
-  }, [inView]);
+  const { profileCards, loading, ref } = useWikiInfiniteScroll(searchTerm);
 
   const hasSearchedProfile = profileCards.some(
     profileCard => profileCard.name.trim().toLowerCase() === searchTerm.trim().toLowerCase(),
